requirements.txt
pylint
pyusb==1.0.0